pyproject.toml
[project]
name = "zchat"
dynamic = ["version"]
description = "Multi-agent collaboration over IRC"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "segno>=1.6.0",
    "tomli-w>=1.0.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.9.0",
    "zchat-protocol>=0.1.0",
    "zchat-channel-server>=0.2.0",
]
license = "MIT"
authors = [{ name = "ezagent42" }]

[project.urls]
Homepage = "https://github.com/ezagent42/zchat"
Repository = "https://github.com/ezagent42/zchat"

[project.scripts]
zchat = "zchat.cli.app:app"

[tool.uv.sources]
zchat-protocol = { path = "zchat-protocol", editable = true }
zchat-channel-server = { path = "zchat-channel-server", editable = true }

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options = { local_scheme = "no-local-version" }

[tool.hatch.build.hooks.vcs]
version-file = "zchat/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["zchat"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-order>=1.3.0",
    "pytest-timeout>=2.4.0",
]
